add user validation

+++ post_restapi/controllers/feed.js
@@ -4,6 +4,54 @@ const path = require('path');
 const { validationResult } = require('express-validator');
 
 const Post = require('../models/post');
+const User = require('../models/user');
+
+exports.getUserStatus = (req, res, next) => {
+    User.findById(req.userId)
+    .then(user => {
+        if(!user){
+            const error = new Error('User not found.');
+            error.statusCode = 404;
+            throw error;
+        }
+        res.status(200).json({
+            message: 'Fetched user status successfully.',
+            status: user.status
+        });
+    })
+    .catch(err => {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    });
+}
+
+exports.updateUserStatus = (req, res, next) => {
+    const status = req.body.status;
+    User.findById(req.userId)
+    .then(user => {
+        if(!user){
+            const error = new Error('User not found.');
+            error.statusCode = 404;
+            throw error;
+        }
+        user.status = status;
+        return user.save();
+    })
+    .then(result => {
+        res.status(200).json({
+            message: 'Status updated successfully!',
+            status: status
+        });
+    })
+    .catch(err => {
+        if (!err.statusCode) {
+            err.statusCode = 500;
+        }
+        next(err);
+    });
+}
 
 exports.getPosts = (req, res, next) => {
     const currentPage = req.query.page || 1;
@@ -52,21 +100,28 @@ exports.createPost = (req, res, next) => {
     const imageUrl = req.file.path;
     const title = req.body.title;
     const content = req.body.content;
+    let creator;
     const post = new Post({
         title: title,
         content: content,
         imageUrl: imageUrl,
-        creator: {
-            name: 'Eunmi'
-        }
+        creator: req.userId
     });
     post
         .save()
         .then(result => {
-            // Create post in db
+            return User.findById(req.userId)
+        })
+        .then(user => {
+            creator = user;
+            user.posts.push(post)
+            return user.save();
+        })
+        .then(result => {
             res.status(201).json({ // 201: success in creating resource
                 message: 'Post created successfully!',
-                post: result
+                post: post,
+                creator: { _id: creator._id, name: creator.name }
             });
         })
         .catch(err => {
@@ -128,6 +183,11 @@ exports.updatePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             if (imageUrl !== post.imageUrl) { // update different image -> delete previous image
                 clearImage(post.imageUrl);
             }
@@ -156,9 +216,21 @@ exports.deletePost = (req, res, next) => {
                 error.statusCode = 404;
                 throw error;
             }
+            if (post.creator.toString !== req.userId) {
+                const error = new Error('Not authorized!');
+                error.statusCode = 403;
+                throw error;
+            }
             // Check logged in user
             clearImage(post.imageUrl);
             return Post.findByIdAndDelete(postId);
+        })
+        .then(result => {
+            return User.findById(req.userId);
+        })
+        .then(user => {
+            user.posts.pull(postId);
+            return user.save();
         })
         .then(result => {
             res.status(200).json({ message: 'Deleted Post' });
